LM api (#30)

+++ modules/streaming/src/main/scala/fi/spectrum/streaming/domain/BlockEvent.scala
@@ -7,13 +7,13 @@ import fi.spectrum.core.domain.BlockId
 @derive(decoder)
 sealed trait BlockEvent {
   val timestamp: Long
-  val height: Long
+  val height: Int
   val id: BlockId
 }
 
 object BlockEvent {
   @derive(decoder)
-  final case class BlockApply(timestamp: Long, height: Long, id: BlockId) extends BlockEvent
+  final case class BlockApply(timestamp: Long, height: Int, id: BlockId) extends BlockEvent
   @derive(decoder)
-  final case class BlockUnapply(timestamp: Long, height: Long, id: BlockId) extends BlockEvent
+  final case class BlockUnapply(timestamp: Long, height: Int, id: BlockId) extends BlockEvent
 }
